requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.32
fredapi>=0.5.1
pandas-market-calendars>=4.3.0
python-dotenv>=1.0.0
pyarrow>=13.0.0
matplotlib>=3.7.0
